Modified printfint 3

+++ _printfint_double.c
@@ -34,7 +34,7 @@ int _printfint(long int i, int j)
  * Return: j
  **/
 
-int _printfint_double(char c, int i, int j)
+int _printfint_double(char c, long int i, int j)
 {
 	if (c == 'i' || c == 'd')
 	{
